requirements.txt
alabaster==0.7.16
arrow==1.3.0
asgiref==3.8.1
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
diff-match-patch==20230430
Django==5.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-import-export==4.1.1
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
git-filter-repo==2.45.0
idna==3.10
imagesize==1.4.1
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
rq==1.16.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-autodoc-typehints==2.4.4
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
tablib==3.5.0
times==0.7
types-python-dateutil==2.9.0.20240906
tzdata==2024.1
urllib3==2.2.3
